requirements.txt
fastapi==0.116.1
passlib==1.7.4
pydantic==2.11.7
pytest==8.4.1
python_jose==3.5.0
SQLAlchemy==2.0.41
starlette==0.47.2
pydantic_settings==2.10.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
python-dotenv